feat: make collectors autonomous and toolbar dynamic
- Move each collector to its own module directory
- Add renderTab() and renderPanel() to CollectorInterface
- Add CollectorAwareInterface for post-instantiation hooks
- Auto-discover and register *Collector.php via ProfilerModule scanner
- Toolbar now iterates collectors dynamically, no hardcoded references
- Rework toolbar UI: white/black theme, improved readability
- Fix toolbar toggle JS: NEO brand always visible, tabs togglable
