+++ neo/Core/Profiler/Collector/CollectorInterface.php
@@ -11,4 +11,10 @@ interface CollectorInterface
      * @return array<string, mixed>
      */
     public function collect(): array;
+
+    /** @param array<string, mixed> $data */
+    public function renderTab(array $data): string;
+
+    /** @param array<string, mixed> $data */
+    public function renderPanel(array $data): string;
 }
