+++ neo/Core/Profiler/CollectorAwareInterface.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace Neo\Core\Profiler;
+
+use Neo\Core\DI\Container;
+
+interface CollectorAwareInterface
+{
+    public function boot(Container $container): void;
+}
